+++ src/dependencies.rs
@@ -15,8 +15,9 @@ use crate::{
     build_manager::{Build, BuildManager},
     custom_flags::Flag,
     per_test_config::TestConfig,
+    status_emitter::RevisionStyle,
     test_result::Errored,
-    CommandBuilder, Config, OutputConflictHandling,
+    CommandBuilder, Config, OutputConflictHandling, TestOk,
 };
 
 #[derive(Default, Debug)]
@@ -390,13 +391,23 @@ impl Flag for DependencyBuilder {
         config: &TestConfig<'_>,
         build_manager: &BuildManager<'_>,
     ) -> Result<(), Errored> {
-        config
-            .status
-            .update_status("waiting for dependencies to finish building".into());
-        let extra_args = build_manager.build(self.clone())?;
-        cmd.args(extra_args);
-        config.status.update_status(String::new());
-        Ok(())
+        let status = config.status.for_revision(
+            "waiting for dependencies to finish building",
+            RevisionStyle::Quiet,
+        );
+        match build_manager.build(self.clone()) {
+            Ok(extra_args) => {
+                cmd.args(extra_args);
+                status.done(&Ok(TestOk::Ok));
+                Ok(())
+            }
+            Err(err) => {
+                let err = Err(err);
+                status.done(&err);
+                #[allow(clippy::unnecessary_literal_unwrap)]
+                Err(err.unwrap_err())
+            }
+        }
     }
 }